add function display product group by brand, add usecase product management

+++ database/migrations/2018_11_01_180147_create_product_details_table.php
@@ -24,7 +24,7 @@ class CreateProductDetailsTable extends Migration
             $table->integer('ram')->nullable();
             $table->integer('internal_memory')->nullable();
             $table->integer('external_memory_card')->nullable();
-            $table->integer('bluetooth')->nullable();
+            $table->float('bluetooth')->nullable();
             $table->float('length')->nullable();
             $table->float('width')->nullable();
             $table->float('thickness')->nullable();
